requirements.txt
streamlit
faiss-cpu
pandas
numpy
docx2txt
pdfminer.six
sentence-transformers
